requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.34.0
django-appconf==1.0.2
django-avatar==4.0.1
django-bower==5.2.0
idna==2.6
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
xlrd==1.1.0
django_cleanup==1.0.1
djangorestframework == 3.7.3